Changes to testing - simplification
